requirements.txt
asgiref==3.3.4
Django==3.2.2
django-tinymce==3.3.0
django-tinymce4-lite==1.8.0
jsmin==2.2.2
pyenchant==3.2.0
pytz==2021.1
sqlparse==0.4.1
